# swagger.json
{
	"components": {
...
		"schemas": {
...
			"EConversionStatus": {
				"enum": [
					"converted",
					"erroneous",
					"in queue",
					"processing"
				],
				"type": "string"
			},
			"IConversionStatus": {
				"properties": {
					"conversionId": {
						"type": "string"
					},
					"sourceFormat": {
						"type": "string"
					},
					"targetFormat": {
						"type": "string"
					},
					"path": {
						"type": "string"
					},
					"retries": {
						"type": "number",
						"format": "double"
					},
					"status": {
						"$ref": "#/components/schemas/EConversionStatus"
					}
				},
				"required": [
					"conversionId",
					"sourceFormat",
					"targetFormat",
					"path",
					"retries",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "conversion-service",
		"version": "0.1.9",
		"description": "A webservice with a REST Api which can convert document/media files using different binaries",
		"license": {
			"name": "GPL-3.0-or-later"
		},
		"contact": {
			"email": "sven.patrick.meier@team-parallax.com",
			"name": "Sven Patrick Meier",
			"url": "https://team-parallax.com/svenpatrickmeier"
		}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/conversion/{conversionId}": {
			"get": {
				"operationId": "GetConvertedFile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IConversionStatus"
								}
							}
						}
					}
				},
				"description": "Returns the current status for a conversion given a conversionId",
				"tags": [
					"Conversion"
				],
				"security": [],
				"parameters": [
					{
						"description": "Unique identifier for the conversion of a file.",
						"in": "path",
						"name": "conversionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/conversion/{conversionId}/download": {
			"get": {
				"operationId": "GetConvertedFileDownload",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							},
							"application/pdf": {
								"schema": {
									"type": "string",
									"format": "byte"
								}
							}
						}
					}
				},
				"description": "Returns the current status for a conversion given a conversionId\nIf status is 'converted' the resulting converted file\nwill be available to download.",
				"tags": [
					"Conversion"
				],
				"security": [],
				"parameters": [
					{
						"description": "the id of the file-conversion that is requested",
						"in": "path",
						"name": "conversionId",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"in": "query",
						"name": "extension",
						"required": false,
						"schema": {
							"default": "mp3",
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "http://localhost:3000/"
		}
	]
}